Update and rename PFL.py to FedAda.py
Federated Learning Adaptive Aggregation (FedAda): The local  models of the devices are aggregated by the server using a dynamic weighted approach based on local model accuracy. accelerate the convergence of the global model. Experiments show that FedAda can not only improve global model generalization and performance but also accelerate the convergence of the global model.